Make GtkBackend build on Windows
Two portability defects in Sources/Gtk, plus the tooling needed to get a GTK 4
in front of the compiler at all. WinUIBackend is untouched and stays the
baseline.

The defects are both Linux/Windows differences in how C types import, and
neither needed a platform conditional:

  gulong is 64 bits on Linux and macOS and 32 on Windows, which is LLP64.
  connectSignal converted it to UInt on the way out, so nothing that takes a
  handler id back -- disconnect, block, unblock -- would compile. The id now
  stays gulong from the call to the store to the disconnect.

  gsize imports as UInt on Linux and UInt64 on Windows. Same width, different
  nominal types, so g_bytes_new rejected the argument. Converted explicitly.

Getting there took three environment findings worth keeping:

  The gvsbuild bundle is not relocatable. 301 of its 302 .pc files hardcode the
  build machine's prefix, and all of them are CRLF.

  SwiftPM parses .pc files itself and splits keyword lines on the first colon,
  so `prefix=C:/gtk4` becomes the keyword `prefix=C` and the variable is never
  defined -- it reports `Expected a value for variable 'prefix'`. Any Windows
  drive letter in a variable value hits this. The installer rewrites prefixes to
  the colon-free `${pcfiledir}/../..` form instead.

  SwiftPM does not apply a systemLibrary's pkgConfig cflags on Windows at all.
  The clang invocation for GtkCHelpers carried only its own include directory,
  with PKG_CONFIG_PATH set and pkg-config reporting the right flags. They have
  to be passed through -Xcc, which the installer prints a command for.

testapp/install_gtk4_windows.zsh does the install and the rewriting.
Acknowledgements/gvsbuild records where the bundle comes from, why MSVC rather
than MSYS2, and that nothing is vendored here.

Verified: `swift build --target GtkBackend` exits 0 on Windows, and Gtk and
GtkBackend still build in WSL. Runtime behaviour on Windows is not verified yet;
testapp/plan/plan-windows-gtk-backend.md sets out the comparison to run.

+++ Sources/Gtk/GObject.swift
@@ -27,7 +27,9 @@ open class GObject: GObjectRepresentable {
         g_object_unref(gobjectPointer)
     }
 
-    private var signals: [(UInt, Any)] = []
+    // `gulong`, not `UInt`: they differ on Windows, where `long` is 32 bits.
+    // See the note in Callbacks/Signals.swift.
+    private var signals: [(gulong, Any)] = []
 
     /// GObject signals sometimes get invoked when you programmatically set
     /// something. If you don't want them to, you can temporarily disable
@@ -44,7 +46,7 @@ open class GObject: GObjectRepresentable {
 
     /// Stores the signal handler ID of the handler that we have registered for
     /// a given signal name.
-    private var blockableSignalIDs: [String: UInt] = [:]
+    private var blockableSignalIDs: [String: gulong] = [:]
 
     open func registerSignals() {}
 
@@ -173,7 +175,7 @@ open class GObject: GObjectRepresentable {
         storeHandler(handlerId, box: box, for: name)
     }
 
-    private func storeHandler(_ id: UInt, box: Any, for signalName: String) {
+    private func storeHandler(_ id: gulong, box: Any, for signalName: String) {
         signals.append((id, box))
         if Self.blockableSignalNames.contains(signalName) {
             blockableSignalIDs[signalName] = id
